requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.2.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
ollama==0.4.4
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
orjson==3.10.12
packaging==24.2
pillow==10.2.0
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
semantic-version==2.10.0
setuptools==75.6.0
setuptools-rust==1.10.2
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
